otlp_protos::opentelemetry::proto::metrics::v1

Struct ExponentialHistogramDataPoint

pub struct ExponentialHistogramDataPoint {

    pub attributes: Vec<KeyValue>,
    pub start_time_unix_nano: u64,
    pub time_unix_nano: u64,
    pub count: u64,
    pub sum: Option<f64>,
    pub scale: i32,
    pub zero_count: u64,
    pub positive: Option<Buckets>,
    pub negative: Option<Buckets>,
    pub flags: u32,
    pub exemplars: Vec<Exemplar>,
    pub min: Option<f64>,
    pub max: Option<f64>,
    pub zero_threshold: f64,

}

ExponentialHistogramDataPoint is a single data point in a timeseries that describes the time-varying values of a ExponentialHistogram of double values. A ExponentialHistogram contains summary statistics for a population of values, it may optionally contain the distribution of those values across a set of buckets.

Fields
attributes: Vec<KeyValue>

The set of key/value pairs that uniquely identify the timeseries from where this point belongs. The list may be empty (may contain 0 elements). Attribute keys MUST be unique (it is not allowed to have more than one attribute with the same key).

start_time_unix_nano: u64

StartTimeUnixNano is optional but strongly encouraged, see the the detailed comments above Metric.

Value is UNIX Epoch time in nanoseconds since 00:00:00 UTC on 1 January 1970.

time_unix_nano: u64

TimeUnixNano is required, see the detailed comments above Metric.

Value is UNIX Epoch time in nanoseconds since 00:00:00 UTC on 1 January 1970.

count: u64

count is the number of values in the population. Must be non-negative. This value must be equal to the sum of the “bucket_counts” values in the positive and negative Buckets plus the “zero_count” field.

sum: Option<f64>

sum of the values in the population. If count is zero then this field must be zero.

Note: Sum should only be filled out when measuring non-negative discrete events, and is assumed to be monotonic over the values of these events. Negative events can be recorded, but sum should not be filled out when doing so. This is specifically to enforce compatibility w/ OpenMetrics, see: https://github.com/prometheus/OpenMetrics/blob/v1.0.0/specification/OpenMetrics.md#histogram

scale: i32

scale describes the resolution of the histogram. Boundaries are located at powers of the base, where:

base = (2^(2^-scale))

The histogram bucket identified by index, a signed integer, contains values that are greater than (base^index) and less than or equal to (base^(index+1)).

The positive and negative ranges of the histogram are expressed separately. Negative values are mapped by their absolute value into the negative range using the same scale as the positive range.

scale is not restricted by the protocol, as the permissible values depend on the range of the data.

zero_count: u64

zero_count is the count of values that are either exactly zero or within the region considered zero by the instrumentation at the tolerated degree of precision. This bucket stores values that cannot be expressed using the standard exponential formula as well as values that have been rounded to zero.

Implementations MAY consider the zero bucket to have probability mass equal to (zero_count / count).

positive: Option<Buckets>

positive carries the positive range of exponential bucket counts.

negative: Option<Buckets>

negative carries the negative range of exponential bucket counts.

flags: u32

Flags that apply to this specific data point. See DataPointFlags for the available flags and their meaning.

exemplars: Vec<Exemplar>

(Optional) List of exemplars collected from measurements that were used to form the data point

min: Option<f64>

min is the minimum value over (start_time, end_time].

max: Option<f64>

max is the maximum value over (start_time, end_time].

zero_threshold: f64

ZeroThreshold may be optionally set to convey the width of the zero region. Where the zero region is defined as the closed interval [-ZeroThreshold, ZeroThreshold]. When ZeroThreshold is 0, zero count bucket stores values that cannot be expressed using the standard exponential formula as well as values that have been rounded to zero.

Implementations

impl ExponentialHistogramDataPoint

pub fn sum(&self) -> f64

Returns the value of sum, or the default value if sum is unset.

pub fn min(&self) -> f64

Returns the value of min, or the default value if min is unset.

pub fn max(&self) -> f64

Returns the value of max, or the default value if max is unset.